requirements.txt
# 深度学习框架
tensorflow>=2.10.0
keras>=2.10.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 数据处理和科学计算
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
tqdm>=4.62.0
jupyter>=1.0.0